add layout and start working on auth again on graphql
