+++ web-app/src/components/layouts/Layout.js
@@ -1,16 +1,25 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import { makeStyles } from "@material-ui/styles";
+import { makeStyles, Container, Box } from "@material-ui/core";
+
+// Components
+import Navbar from "../navbars/Navbar";
+import Footer from "./Footer";
 
 const useStyles = makeStyles((theme) => ({
   root: {
     display: "flex",
   },
-  mainContent: {
-    objectFit: "contain",
-    height: "100vh",
+  appBarSpacer: theme.mixins.toolbar,
+  content: {
     flexGrow: 1,
+    height: "100vh",
+    overflow: "auto",
+  },
+  container: {
+    paddingTop: theme.spacing(4),
+    paddingBottom: theme.spacing(4),
   },
 }));
 
@@ -19,7 +28,16 @@ const Layout = ({ children }) => {
 
   return (
     <div className={classes.root}>
-      <main className={classes.mainContent}>{children}</main>
+      <Navbar />
+      <main className={classes.content}>
+        <div className={classes.appBarSpacer} />
+        <Container maxWidth="lg" className={classes.container}>
+          {children}
+          <Box pt={4}>
+            <Footer />
+          </Box>
+        </Container>
+      </main>
     </div>
   );
 };
