--- web-app/src/constants/Routes.js
@@ -1,9 +1,0 @@
-const ROUTES = Object.freeze({
-  ROOT: "/",
-  LOGIN: "/login",
-  REGISTER: "/register",
-  DASH_BOARD: "/dashboard",
-  CARDS: "/cards",
-});
-
-export default ROUTES;
