+++ web-app/src/constants/uiConfigs.js
@@ -1,0 +1,5 @@
+const uiConfigs = Object.freeze({
+  drawerWidth: 240,
+});
+
+export default uiConfigs;
