+++ web-app/src/graphql/mutations/LOGIN_USER.js
@@ -5,6 +5,7 @@ const LOGIN_USER = gql`
     loginUser(input: $input) {
       name
       email
+      jwtToken
     }
   }
 `;
