+++ web-app/src/graphql/queries/GET_USER.js
@@ -1,0 +1,10 @@
+import { gql } from "@apollo/client";
+
+export default gql`
+  query GET_USER {
+    getUser {
+      name
+      email
+    }
+  }
+`;
