+++ web-server/package.json
@@ -11,6 +11,7 @@
     "apollo-server": "^2.21.0",
     "apollo-server-express": "^2.21.0",
     "bcryptjs": "^2.4.3",
+    "chalk": "^4.1.0",
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
     "express": "^4.17.1",
@@ -31,8 +32,8 @@
     "prettier": "2.2.1"
   },
   "scripts": {
-    "start": "node src/index.js",
-    "start:dev": "nodemon src/index.js --ignore web-app",
+    "start": "node src/index.js --colors",
+    "start:dev": "nodemon src/index.js --ignore web-app --colors",
     "test": "jest --coverage --watchAll",
     "test:ci": "CI=true jest --coverage --watchAll=false"
   }
