+++ web-server/src/configs/serverConfigs.js
@@ -10,6 +10,7 @@ const serverConfigs = Object.freeze({
   PORT: process.env.PORT || 5000,
   MONGO_URI: process.env.MONGO_URI || keys.MONGO_URI,
   SECRET_KEY: process.env.SECRET_KEY || keys.secretOrKey,
+  COOKIE_NAME: "secretToken",
 });
 
 module.exports = { serverConfigs };
