+++ web-server/src/graphql/index.js
@@ -4,7 +4,6 @@ const resolvers = require("./resolvers");
 
 // The ApolloServer constructor requires two parameters: your schema
 // definition and your set of resolvers.
-const server = new ApolloServer({ typeDefs, resolvers });
 
 module.exports = new ApolloServer({
   typeDefs,
