+++ web-server/src/graphql/resolvers/utils/cryptoHelpers.js
@@ -1,0 +1,6 @@
+const crypto = require("crypto");
+
+const generateToken = (length) =>
+  crypto.randomBytes(length || 32).toString("hex");
+
+module.exports = { generateToken };
